theme: create shortcode for mailing list inclusion
The `[bsplist product="product name"]` shortcode inserts a mailing list
sign-up form into the page, so no more crazy code hacking is required,
and easy to update code across the site when needed.

Update is done by replacing the code in `partials/mailchimp-form.php`.
At the moment the update process is the following:

1. Log into Mailchimp to the VIA mailing lists
2. Select the "VIA BSP Notifications" list
3. Select "Signup forms" / "Embedded forms" / "Naked"
4. Options for generated HTML code:
  1. Change form title as desired (eg. see current one)
  2. Check "Show only required fields" option
  3. Check "Show interest group fields" option
  4. Uncheck "Show required fields indicators" option

Copy paste the resulting HTML into the file mentioned above, replacing
the existing content.

The shortcode also includes as cript to pre-select items:
the "product name" variable should be equal to the label of the desired field.